Refactor Indexer.Fetcher.TokenInstance naming
