+++ apps/indexer/lib/indexer/fetcher/token_instance/helper.ex
@@ -1,4 +1,4 @@
-defmodule Indexer.Fetcher.TokenInstance do
+defmodule Indexer.Fetcher.TokenInstance.Helper do
   @moduledoc """
     Common functions for Indexer.Fetcher.TokenInstance fetchers
   """
