+++ apps/indexer/lib/indexer/fetcher/token_instance/realtime.ex
@@ -1,4 +1,4 @@
-defmodule Indexer.Fetcher.Retry.TokenInstance do
+defmodule Indexer.Fetcher.TokenInstance.Realtime do
   @moduledoc """
   Fetches information about a token instance.
   """
@@ -6,7 +6,7 @@ defmodule Indexer.Fetcher.Retry.TokenInstance do
   use Indexer.Fetcher, restart: :permanent
   use Spandex.Decorators
 
-  import Indexer.Fetcher.TokenInstance
+  import Indexer.Fetcher.TokenInstance.Helper
 
   alias Explorer.Chain
   alias Indexer.BufferedTask
@@ -35,34 +35,56 @@ defmodule Indexer.Fetcher.Retry.TokenInstance do
   end
 
   @impl BufferedTask
-  def init(initial_acc, reducer, _) do
-    {:ok, acc} =
-      Chain.stream_token_instances_with_error(initial_acc, fn data, acc ->
-        reducer.(data, acc)
-      end)
-
-    acc
+  def init(_, _, _) do
+    {0, []}
   end
 
   @impl BufferedTask
-  def run([%{contract_address_hash: hash, token_id: token_id, updated_at: updated_at}], _json_rpc_named_arguments) do
-    refetch_interval = Application.get_env(:indexer, __MODULE__)[:refetch_interval]
-
-    if updated_at
-       |> DateTime.add(refetch_interval, :millisecond)
-       |> DateTime.compare(DateTime.utc_now()) != :gt do
-      fetch_instance(hash, token_id, true)
+  def run([%{contract_address_hash: hash, token_id: token_id}], _json_rpc_named_arguments) do
+    if not Chain.token_instance_exists?(token_id, hash) do
+      fetch_instance(hash, token_id, false)
     end
 
     :ok
   end
 
+  @doc """
+  Fetches token instance data asynchronously.
+  """
+  def async_fetch(data) do
+    async_fetch(data, __MODULE__.Supervisor.disabled?())
+  end
+
+  def async_fetch(_data, true), do: :ok
+
+  def async_fetch(token_transfers, _disabled?) when is_list(token_transfers) do
+    data =
+      token_transfers
+      |> Enum.reject(fn token_transfer -> is_nil(token_transfer.token_ids) end)
+      |> Enum.map(fn token_transfer ->
+        Enum.map(token_transfer.token_ids, fn token_id ->
+          %{
+            contract_address_hash: token_transfer.token_contract_address_hash,
+            token_id: token_id
+          }
+        end)
+      end)
+      |> List.flatten()
+      |> Enum.uniq()
+
+    BufferedTask.buffer(__MODULE__, data)
+  end
+
+  def async_fetch(data, _disabled?) do
+    BufferedTask.buffer(__MODULE__, data)
+  end
+
   defp defaults do
     [
-      flush_interval: :timer.minutes(10),
+      flush_interval: 100,
       max_concurrency: Application.get_env(:indexer, __MODULE__)[:concurrency] || @default_max_concurrency,
       max_batch_size: @default_max_batch_size,
-      poll: true,
+      poll: false,
       task_supervisor: __MODULE__.TaskSupervisor
     ]
   end
